fix: showing if the var is false on first init
Example if the variable set to false at first initialization.. the div supposed to vanish on inspect element at first load.. so it acts like a ngif. .. because the old one only triggering the vanishing if user change the variable to false on the fly.. regardless on the variable first initialization.